add image_files for products - hosted on AWS-S3

+++ lib/tasks/db.rake
@@ -1,9 +1,15 @@
 namespace :db do
   desc "run db:drop, db:create, db:migrate and db:seed tasks in a row."
   task :rebuild => :environment do
-    system "rake db:drop"
-    system "rake db:create"
-    system "rake db:migrate"
-    system "rake db:seed"
+    if Rails.env.production?
+      system "heroku pg:reset DATABASE_URL --confirm quiet-mountain-4649"
+      system "heroku run rake db:migrate"
+      system "heroku run rake db:seed"
+    else
+      system "rake db:drop"
+      system "rake db:create"
+      system "rake db:migrate"
+      system "rake db:seed"
+    end
   end
 end
